+++ demos/compose/src/main/java/androidx/media3/demo/compose/MainActivity.kt
@@ -30,6 +30,8 @@ import androidx.media3.common.MediaItem
 import androidx.media3.common.Player
 import androidx.media3.demo.compose.data.videos
 import androidx.media3.exoplayer.ExoPlayer
+import androidx.media3.ui.compose.PlayerSurface
+import androidx.media3.ui.compose.SURFACE_TYPE_SURFACE_VIEW
 
 class MainActivity : ComponentActivity() {